requirements.txt
folium==0.14.0
asgiref==3.6.0
Django==4.1.4
django-ckeditor==6.5.1
django-decouple==2.0
Khayyam==3.0.17
Pillow==9.3.0
django-widget-tweaks==1.4.12
geocoder==1.38.1
django-admin-thumbnails==0.2.6
django-allauth
social-auth-app-django==5.0.0
social-auth-core==4.2.0
markdown2==2.4.3
